Moved ProductGrabber into its own module
